\name{QualityGraphsOligo-package}
\alias{QualityGraphsOligo-package}
\alias{QualityGraphsOligo}
\docType{package}
\title{
\packageTitle{QualityGraphsOligo}
}
\description{
\packageDescription{QualityGraphsOligo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{QualityGraphsOligo}
\packageIndices{QualityGraphsOligo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{QualityGraphsOligo}

Maintainer: \packageMaintainer{QualityGraphsOligo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
